rename sp3 local module
