+++ rinex-cli/src/plot/record/mod.rs
@@ -2,10 +2,10 @@ mod ionex;
 mod meteo;
 mod navigation;
 mod observation;
-mod sp3;
+mod sp3_plot;
 
 pub use ionex::plot_tec_map;
 pub use meteo::plot_meteo;
 pub use navigation::plot_navigation;
 pub use observation::plot_observation;
-pub use sp3::plot_residual_ephemeris;
+pub use sp3_plot::plot_residual_ephemeris;
