+++ rinex-cli/src/plot/record/sp3_plot.rs
